Converted the Shop menu to use Fleets&misc changes
The Shop menu does not fully function yet: a custom to_string function
that only leaves 1 decimal must be used. Perhaps utility.h already has
one.

The string 'blank' was removed because it's a legacy thing from 2010.
More #pragma ignores were removed.
A lot of type casting was added.

+++ ChartedWaters/State_shipPartInventory.cpp
@@ -20,7 +20,7 @@ bool State_shipPartInventory::Init()
 
   redrawLeft(partList);
   ShipPart* part = 0;
-  if (selector >= 3 && selector - 3 < partList.size())
+  if (selector >= 3 && (size_t)selector - 3 < partList.size())
     part = partList[selector-3];
   redrawRight(part);
 
@@ -34,7 +34,7 @@ void State_shipPartInventory::Update()
       redrawLeft(partList);
 
       ShipPart* part = 0;
-      if (selector >= 3 && selector - 3 < partList.size())
+      if (selector >= 3 && (size_t)selector - 3 < partList.size())
         part = partList[selector-3];
       redrawRight(part);
